chore(move): move `internal/macaroon` to `api/macaroon`
jaas was using MacaroonNamespace from "github.com/juju/juju/core/macaroon" in juju 3, it got moved
to internal in juju 4. I'm moving it back to `api`
