+++ api/macaroon/store_test.go
@@ -13,11 +13,11 @@ import (
 	"github.com/juju/tc"
 	"gopkg.in/macaroon.v2"
 
+	apimacaroon "github.com/juju/juju/api/macaroon"
 	domainmacaroon "github.com/juju/juju/domain/macaroon"
 	macaroonservice "github.com/juju/juju/domain/macaroon/service"
 	macaroonstate "github.com/juju/juju/domain/macaroon/state"
 	domaintesting "github.com/juju/juju/internal/changestream/testing"
-	internalmacaroon "github.com/juju/juju/internal/macaroon"
 )
 
 type storeSuite struct {
@@ -50,7 +50,7 @@ var policyMinute = dbrootkeystore.Policy{
 }
 
 func (s *storeSuite) TestNewRootKeyStore(c *tc.C) {
-	expireableStorage := internalmacaroon.NewRootKeyStore(s.macaroonService, policyMinute, s.clock)
+	expireableStorage := apimacaroon.NewRootKeyStore(s.macaroonService, policyMinute, s.clock)
 
 	key1, id, err := expireableStorage.RootKey(c.Context())
 	c.Assert(err, tc.ErrorIsNil)
@@ -69,7 +69,7 @@ func (s *storeSuite) TestExpiredRootKeyRemovedGet(c *tc.C) {
 	})
 	c.Assert(err, tc.ErrorIsNil)
 
-	expireableStorage := internalmacaroon.NewRootKeyStore(s.macaroonService, policyMinute, s.clock)
+	expireableStorage := apimacaroon.NewRootKeyStore(s.macaroonService, policyMinute, s.clock)
 	_, err = expireableStorage.Get(c.Context(), []byte("id"))
 	c.Assert(err, tc.ErrorIs, bakery.ErrNotFound)
 
@@ -86,7 +86,7 @@ func (s *storeSuite) TestExpiredRootKeyRemovedRootKey(c *tc.C) {
 	})
 	c.Assert(err, tc.ErrorIsNil)
 
-	expireableStorage := internalmacaroon.NewRootKeyStore(s.macaroonService, policyMinute, s.clock)
+	expireableStorage := apimacaroon.NewRootKeyStore(s.macaroonService, policyMinute, s.clock)
 	_, _, err = expireableStorage.RootKey(c.Context())
 	c.Assert(err, tc.ErrorIsNil)
 
@@ -95,7 +95,7 @@ func (s *storeSuite) TestExpiredRootKeyRemovedRootKey(c *tc.C) {
 }
 
 func (s *storeSuite) TestCheckNewMacaroon(c *tc.C) {
-	expireableStorage := internalmacaroon.NewRootKeyStore(s.macaroonService, policyMinute, s.clock)
+	expireableStorage := apimacaroon.NewRootKeyStore(s.macaroonService, policyMinute, s.clock)
 
 	b := bakery.New(bakery.BakeryParams{
 		RootKeyStore: expireableStorage,
@@ -112,7 +112,7 @@ func (s *storeSuite) TestCheckNewMacaroon(c *tc.C) {
 }
 
 func (s *storeSuite) TestExpiryTime(c *tc.C) {
-	expireableStorage := internalmacaroon.NewRootKeyStore(s.macaroonService, dbrootkeystore.Policy{
+	expireableStorage := apimacaroon.NewRootKeyStore(s.macaroonService, dbrootkeystore.Policy{
 		ExpiryDuration: 5 * time.Millisecond,
 	}, s.clock)
 
@@ -250,7 +250,7 @@ func (s *storeSuite) testRootKeyUsesKeysValidWithPolicy(c *tc.C, i int) {
 	err := s.macaroonService.InsertKeyContext(c.Context(), test.key)
 	c.Assert(err, tc.ErrorIsNil, tc.Commentf(test.about))
 
-	store := internalmacaroon.NewRootKeyStore(s.macaroonService, test.policy, s.clock)
+	store := apimacaroon.NewRootKeyStore(s.macaroonService, test.policy, s.clock)
 	key, id, err := store.RootKey(c.Context())
 	c.Assert(err, tc.ErrorIsNil, tc.Commentf(test.about))
 	if test.expect {
@@ -269,7 +269,7 @@ func (s *storeSuite) TestRootKey(c *tc.C) {
 	policy := dbrootkeystore.Policy{
 		ExpiryDuration: 5 * time.Minute,
 	}
-	store := internalmacaroon.NewRootKeyStore(s.macaroonService, policy, s.clock)
+	store := apimacaroon.NewRootKeyStore(s.macaroonService, policy, s.clock)
 	key, id, err := store.RootKey(c.Context())
 	c.Assert(err, tc.ErrorIsNil)
 	c.Assert(key, tc.HasLen, 24)
@@ -284,7 +284,7 @@ func (s *storeSuite) TestRootKey(c *tc.C) {
 	c.Assert(id1, tc.DeepEquals, id)
 
 	// A different store instance should get the same root key.
-	store1 := internalmacaroon.NewRootKeyStore(s.macaroonService, policy, s.clock)
+	store1 := apimacaroon.NewRootKeyStore(s.macaroonService, policy, s.clock)
 	key1, id1, err = store1.RootKey(c.Context())
 	c.Assert(err, tc.ErrorIsNil)
 	c.Assert(key1, tc.DeepEquals, key)
@@ -329,7 +329,7 @@ func (s *storeSuite) TestPreferredRootKeyFromDatabaseLatest(c *tc.C) {
 		c.Assert(err, tc.ErrorIsNil)
 	}
 
-	store := internalmacaroon.NewRootKeyStore(s.macaroonService, dbrootkeystore.Policy{
+	store := apimacaroon.NewRootKeyStore(s.macaroonService, dbrootkeystore.Policy{
 		ExpiryDuration: 7 * time.Minute,
 	}, s.clock)
 	s.now = moment.Add(5 * time.Minute)
@@ -361,7 +361,7 @@ func (s *storeSuite) TestPreferredRootKeyFromDatabaseExclude(c *tc.C) {
 		c.Assert(err, tc.ErrorIsNil)
 	}
 
-	store := internalmacaroon.NewRootKeyStore(s.macaroonService, dbrootkeystore.Policy{
+	store := apimacaroon.NewRootKeyStore(s.macaroonService, dbrootkeystore.Policy{
 		ExpiryDuration: 7 * time.Minute,
 	}, s.clock)
 	s.now = moment.Add(5 * time.Minute)
@@ -371,7 +371,7 @@ func (s *storeSuite) TestPreferredRootKeyFromDatabaseExclude(c *tc.C) {
 }
 
 func (s *storeSuite) TestGet(c *tc.C) {
-	store := internalmacaroon.NewRootKeyStore(s.macaroonService, dbrootkeystore.Policy{
+	store := apimacaroon.NewRootKeyStore(s.macaroonService, dbrootkeystore.Policy{
 		ExpiryDuration: 30 * time.Minute,
 	}, s.clock)
 	type idKey struct {
