+++ apiserver/internal/crossmodel/bakery/jaas_test.go
@@ -18,9 +18,9 @@ import (
 	gomock "go.uber.org/mock/gomock"
 	"gopkg.in/macaroon.v2"
 
+	apimacaroon "github.com/juju/juju/api/macaroon"
 	coreerrors "github.com/juju/juju/core/errors"
 	loggertesting "github.com/juju/juju/internal/logger/testing"
-	internalmacaroon "github.com/juju/juju/internal/macaroon"
 )
 
 type jaasBakerySuite struct {
@@ -34,7 +34,7 @@ func TestJAASBakerySuite(t *testing.T) {
 func (s *jaasBakerySuite) TestNewJAASOfferBakery(c *tc.C) {
 	defer s.setupMocks(c).Finish()
 
-	checker := checkers.New(internalmacaroon.MacaroonNamespace)
+	checker := checkers.New(apimacaroon.MacaroonNamespace)
 	bakery, err := NewJAASOfferBakery(
 		s.keyPair,
 		"juju model",
