+++ cmd/containeragent/initialize/package_test.go
@@ -30,6 +30,7 @@ func (*importSuite) TestImports(c *tc.C) {
 		"api/agent/caasapplication",
 		"api/agent/keyupdater",
 		"api/base",
+		"api/macaroon",
 		"api/common",
 		"api/watcher",
 		"apiserver/errors",
@@ -92,7 +93,6 @@ func (*importSuite) TestImports(c *tc.C) {
 		"internal/featureflag",
 		"internal/http",
 		"internal/logger",
-		"internal/macaroon",
 		"internal/password",
 		"internal/pki",
 		"internal/proxy",
