+++ cmd/jujud-controller/agent/machine/manifolds.go
@@ -24,6 +24,7 @@ import (
 	"github.com/juju/juju/api"
 	"github.com/juju/juju/api/base"
 	"github.com/juju/juju/api/controller/crosscontroller"
+	"github.com/juju/juju/api/macaroon"
 	"github.com/juju/juju/caas"
 	"github.com/juju/juju/cmd/jujud-controller/util"
 	"github.com/juju/juju/core/flightrecorder"
@@ -43,7 +44,6 @@ import (
 	internalhttp "github.com/juju/juju/internal/http"
 	internallease "github.com/juju/juju/internal/lease"
 	internallogger "github.com/juju/juju/internal/logger"
-	"github.com/juju/juju/internal/macaroon"
 	internalobjectstore "github.com/juju/juju/internal/objectstore"
 	proxyconfig "github.com/juju/juju/internal/proxy/config"
 	"github.com/juju/juju/internal/s3client"
